requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
Django==4.2.13
django-crispy-forms==2.1
django-filter==24.2
django-simple-history==3.5.0
idna==3.7
psycopg2-binary==2.9.9
PyYAML==6.0.1
requests==2.31.0
slippers==0.6.2
sqlparse==0.5.0
typeguard==2.13.3
typing_extensions==4.11.0
urllib3==2.2.1
whitenoise==6.6.0
wrapt==1.16.0
